--- l10n_br_account_payment_cobranca/views/account_payment_mode.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="cnab_view_payment_mode_form_inherit_l10n_br_payment"
-            model="ir.ui.view">
-        <field name="name">cnab.payment.mode.l10n_br_payment.form.inherit
-        </field>
-        <field name="model">account.payment.mode</field>
-        <field name="inherit_id"
-               ref="l10n_br_account_payment_order.l10n_br_account_payment_mode_form"/>
-        <field name="priority" eval="40"/>
-        <field name="arch" type="xml">
-            <group name="main" position="after">
-                 <group name="l10n_br_account_payment_taxes" string="Taxas">
-                    <field name="product_tax_id"/>
-                    <field name="tax_account_id"
-                           attrs="{'invisible': [('product_tax_id', '=', False)], 'required': [('product_tax_id', '!=', False)]}"/>
-                </group>
-            </group>
-            <group name="l10n_br_account_payment_order-geral" position="inside">
-                <field name="condition_issuing_paper"/>
-                <field name="cnab_percent_interest"/>
-                <field name="cnab_sequence_id"/>
-                <field name="communication_2"/>
-            </group>
-            <notebook name="l10n_br_account_payment_order" position="inside">
-                <page string="Cobranca">
-                    <group>
-                        <field name="bank_code_bc" invisible="1"/>
-                        <field name="boleto_type"/>
-                        <field name="generate_own_number"/>
-                        <field name="own_number_sequence"/>
-                        <field name="boleto_convetion"/>
-                        <field name="boleto_wallet"/>
-                        <field name="boleto_modality"/>
-                        <field name="boleto_variation"/>
-                        <field name="boleto_cnab_code"/>
-                        <field name="boleto_accept"/>
-                        <field name="boleto_species"/>
-                        <field name="boleto_protest_code"/>
-                        <field name="boleto_days_protest"/>
-                        <field name="boleto_byte_idt" attrs="{'invisible': [('bank_code_bc', 'not in', ('748','756') )]}"/>
-                        <field name="boleto_post" attrs="{'invisible': [('bank_code_bc', 'not in', ('748','756') )]}"/>
-                    </group>
-                </page>
-                <page string="Pagamento">
-                    <group>
-                        <field name="service_type"/>
-                        <field name="release_form"/>
-                        <field name="code_convetion"/>
-                        <field name="doc_finality_code"/>
-                        <field name="ted_finality_code"/>
-                        <field name="complementary_finality_code"/>
-                        <field name="favored_warning"/>
-                    </group>
-                </page>
-                <page string="Contabilidade">
-                    <group>
-                        <field name="default_account_id"/>
-                        <field name="default_tax_account_id"/>
-                    </group>
-                </page>
-            </notebook>
-        </field>
-    </record>
-</odoo>
